requirements.txt
pymavlink>=1.1.36
pyserial>=2.7
pyzmq>=14.3.1
